requirements.txt
fastapi
uvicorn
python-nmap
pydantic
pytest
# scapy (optional, needed for advanced packet work)
# scapy
